Remove version.py
