+++ nilearn/plotting/__init__.py
@@ -6,6 +6,8 @@ import importlib
 import warnings
 
 
+OPTIONAL_MATPLOTLIB_MIN_VERSION = "3.3.0"
+
 ###############################################################################
 # Make sure that we don't get DISPLAY problems when running without X on
 # unices
@@ -20,8 +22,7 @@ def _set_mpl_backend():
             skip_if_running_tests('matplotlib not installed')
         raise
     else:
-        from ..version import (_compare_version,
-                               OPTIONAL_MATPLOTLIB_MIN_VERSION)
+        from .._utils.helpers import _compare_version
         # When matplotlib was successfully imported we need to check
         # that the version is greater that the minimum required one
         mpl_version = getattr(matplotlib, "__version__", "0.0.0")
